made many errors

+++ libft/includes/libft.h
@@ -53,7 +53,7 @@ void				ft_putchar_fd(char c, int fd);
 void				ft_putstr_fd(char const *s, int fd);
 void				ft_putendl_fd(char const *s, int fd);
 void				ft_putnbr_fd(int n, int fd);
-char				*ft_itoa(int n);
+char				*ft_itoa(long long int n);
 char				*ft_strnew(size_t size);
 void				ft_strdel(char **as);
 void				*ft_memalloc(size_t size);
